restructure notification models
