+++ insights_backend/insights_backend/core/models/__init__.py
@@ -1,5 +1,6 @@
-from ...notifications.models import Alert, NotificationChannelConfig, Report
+# from .notifications.models import Alert, NotificationChannelConfig, Report
 from .base import InsightsSchemaBaseModel
+from .notifications import Alert, NotificationChannelConfig, Report
 from .tenant import (
     Tenant,
     TenantConfig,
